Cleanup Code 2.0 (#23)
The next stage of Code Clean-up!

Added a bunch of TODO comments. 

Main thing to fix is to get the predict flow working on main with
MultiPatientEnv. @alyssadsouza @ShivamSingal please check logs of latest
runs to see how the Predict section is failing.

@y-mx can start working on DQN with LSTM to see what functionality we
need to add to our config file processing.

---------

Co-authored-by: alyssadsouza <alyssa.dsouza63@gmail.com>
Co-authored-by: Yimeng Xie <yimeng.xie@cloudsparklabs.com>
Co-authored-by: Christopher Risi <cjrisi@watgpu.cs.uwaterloo.ca>